+++ authority-portal-backend/authority-portal-quarkus/src/test/kotlin/de/sovity/authorityportal/web/thirdparty/keycloak/KeycloakServiceTest.kt
@@ -1,16 +1,17 @@
 package de.sovity.authorityportal.web.thirdparty.keycloak
 
 import de.sovity.authorityportal.web.thirdparty.keycloak.model.KeycloakUserDto
-import de.sovity.authorityportal.web.thirdparty.keycloak.KeycloakService
-import de.sovity.authorityportal.web.thirdparty.keycloak.KeycloakUserMapper
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.keycloak.admin.client.Keycloak
 import org.keycloak.admin.client.resource.RealmResource
+import org.keycloak.admin.client.resource.RoleMappingResource
+import org.keycloak.admin.client.resource.RoleScopeResource
 import org.keycloak.admin.client.resource.UserResource
 import org.keycloak.admin.client.resource.UsersResource
+import org.keycloak.representations.idm.RoleRepresentation
 import org.keycloak.representations.idm.UserRepresentation
 import org.mockito.ArgumentMatchers.eq
 import org.mockito.InjectMocks
@@ -73,6 +74,31 @@ class KeycloakServiceTest {
 
         // assert
         assertThat(actual).isEqualTo(expected)
+    }
+
+    @Test
+    fun testGetUserRoles() {
+        // arrange
+        val userId = "123"
+        val userResource = mock(UserResource::class.java)
+        val roleMappingResource = mock(RoleMappingResource::class.java)
+        val roleScopeResource = mock(RoleScopeResource::class.java)
+        val role1 = mock(RoleRepresentation::class.java)
+        val role2 = mock(RoleRepresentation::class.java)
+        `when`(role1.name).thenReturn("UR_AUTHORITY-PORTAL_AUTHORITY-USER")
+        `when`(role2.name).thenReturn("UR_AUTHORITY-PORTAL_PARTICIPANT-CURATOR")
+
+        `when`(usersResource.get(eq(userId))).thenReturn(userResource)
+        `when`(userResource.roles()).thenReturn(roleMappingResource)
+        `when`(roleMappingResource.realmLevel()).thenReturn(roleScopeResource)
+        `when`(roleScopeResource.listEffective()).thenReturn(listOf(role1, role2))
+
+        // act
+        val userRoles = keycloakService.getUserRoles(userId)
+
+        // assert
+        assert(userRoles.contains("UR_AUTHORITY-PORTAL_AUTHORITY-USER"))
+        assert(userRoles.contains("UR_AUTHORITY-PORTAL_PARTICIPANT-CURATOR"))
     }
 
     private fun mockUserResource(representation: UserRepresentation): UserResource {